Resolved some issues with permutation testing and pvalues
Consider the case where we have 1 calcium event on a given trial.
Shuffling trial types will give a pmf of two points [-1, 1]  for the
shufdelta curves. In the function, we then ask for shufdelta > 1, which
is 0 because the max value for shufdelta is 1, so that gives us an
erroneous p-value of 0. To get around this, I white noise the permuted
delta curves so that half of them actually exceed the empirical delta
curve. This doesn't affect other responses (e.g., 5 calcium events on
one turn vs. 2 on the other) because our mean firing rates are discrete
enough that the white noise shouldn't bring, say, a permuted 4-3 curve
into significance.